requirements.txt
discord.py>=1.5.1
dblpy==0.4.0
psycopg2-binary==2.8.6
